requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
certifi==2022.5.18.1
charset-normalizer==2.0.12
cloudinary==1.29.0
Django==4.0.4
gunicorn==20.1.0
idna==3.3
Pillow==9.1.1
psycopg2-binary==2.9.3
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.9
whitenoise==6.1.0
